requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
supabase>=1.2.0
google-generativeai>=0.8.0
firecrawl-py>=1.0.0
python-dotenv>=1.0.0
httpx>=0.24.1
pydantic>=2.0.0
python-multipart>=0.0.6
